backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
python-dotenv==1.0.0
asyncpg==0.29.0
sqlalchemy==2.0.23
qdrant-client==1.8.0
cohere==4.34
google-generativeai==0.4.0
better-exceptions==0.3.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.0